requirements.txt
numpy>=1.24
Pillow>=10
tensorboard>=2.14
omegaconf>=2.3.0
